YOLOv3/requirements.txt
tensorflow-cpu==1.15.0
keras==2.1.5
matplotlib
Pillow
